pyproject.toml
[project]
name = "eqmd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-allauth>=0.57.0",
    "django-model-utils>=5.0.0",
    "django-taggit>=6.1.0",
    "ffmpeg-python>=0.2.0",
    "pillow>=11.2.1",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "coverage>=7.8.2",
    "django-debug-toolbar>=5.2.0",
    "factory-boy>=3.3.3",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
]
prod = [
    "gunicorn>=23.0.0",
]
